Zone choosing fix
