+++ webui/src/store/actionTypes.js
@@ -99,6 +99,8 @@ export const CLUSTER_SELF_UPDATE = 'CLUSTER_SELF_UPDATE';
 
 export const CLUSTER_PAGE_STATE_RESET = 'CLUSTER_PAGE_STATE_RESET';
 
+export const CLUSTER_PAGE_ZONE_UPDATE = 'CLUSTER_PAGE_ZONE_UPDATE';
+
 export const SET_BOOSTRAP_VSHARD_PANEL_VISIBLE = 'SET_BOOSTRAP_VSHARD_PANEL_VISIBLE';
 export const SET_FAILOVER_MODAL_VISIBLE = 'SET_FAILOVER_MODAL_VISIBLE';
 export const SET_PROBE_SERVER_MODAL_VISIBLE = 'SET_PROBE_SERVER_MODAL_VISIBLE';
