+++ webui/src/store/effector/clusterZones/index.js
@@ -1,0 +1,52 @@
+// @flow
+import {
+  createStore,
+  createEvent,
+  createEffect,
+  type Effect,
+  type Store
+} from 'effector';
+import { editServers } from '../../request/clusterPage.requests';
+import { getErrorMessage } from '../../../api';
+import { clusterPageMount } from '../cluster';
+import store from '../../instance';
+import { CLUSTER_PAGE_ZONE_UPDATE } from '../../actionTypes';
+
+
+export const zoneAddModalOpen = createEvent<string>('zone adding modal open');
+export const zoneAddModalClose = createEvent<mixed>('zone adding modal close');
+
+export const $zoneAddForInstance: Store<string | null> = createStore(null);
+export const $zoneAddModalVisible = $zoneAddForInstance.map<boolean>(uuid => !!uuid);
+export const $error: Store<?string> = createStore(null);
+
+export const submitZoneFx: Effect<
+  { uuid: string | null, zone: string },
+  void,
+  Error
+> = createEffect(
+  'submit server zone change',
+  {
+    handler: async ({ uuid, zone }) => {
+      if (uuid && zone) await editServers([{ uuid, zone }]);
+      else throw new Error('Invalid zone name or UUID');
+    }
+  }
+);
+
+// init
+$zoneAddForInstance
+  .on(zoneAddModalOpen, (_, uuid) => uuid)
+  .reset(submitZoneFx.done)
+  .reset(zoneAddModalClose)
+  .reset(clusterPageMount);
+
+$error
+  .on(submitZoneFx.failData, (_, error) => getErrorMessage(error))
+  .reset(submitZoneFx)
+  .reset(submitZoneFx.done)
+  .reset(zoneAddModalClose)
+  .reset(zoneAddModalOpen)
+  .reset(clusterPageMount);
+
+submitZoneFx.done.watch(() => store.dispatch({ type: CLUSTER_PAGE_ZONE_UPDATE }));
